+++ noodles-vcf/CHANGELOG.md
@@ -17,7 +17,11 @@
 
   * vcf/reader: `Reader::{records, query}` now parse records with the header.
 
-  * vcf/record/format: Prefer key from header.
+  * vcf/record: Move `Format` under genotypes module.
+
+    Use `vcf::record::genotypes::Format` instead.
+
+  * vcf/record/genotypes/format: Prefer key from header.
 
   * vcf/record/info/field: Prefer key from header.